54898 Movido notificação de compensação de Ausência para o Worker Rabbit
