+++ src/SME.SGP.Aplicacao/Interfaces/CasosDeUso/CompensacaoAusencia/INotificarCompensacaoAusenciaUseCase.cs
@@ -1,0 +1,6 @@
+﻿namespace SME.SGP.Aplicacao.Interfaces
+{
+    public interface INotificarCompensacaoAusenciaUseCase : IRabbitUseCase
+    {
+    }
+}
