+++ src/SME.SGP.Dominio.Interfaces/Servicos/IServicoNotificacaoFrequencia.cs
@@ -7,7 +7,6 @@ namespace SME.SGP.Dominio.Interfaces
     {
         Task ExecutaNotificacaoRegistroFrequencia();
         Task VerificaRegraAlteracaoFrequencia(long registroFrequenciaId, DateTime criadoEm, DateTime alteradoEm, long usuarioAlteracaoId);
-        Task NotificarCompensacaoAusencia(long compensacaoId);
         Task NotificarAlunosFaltosos();
         Task NotificarAlunosFaltososBimestre();
     }
