+++ src/SME.SGP.Dominio.Servicos/ServicoCompensacaoAusencia.cs
@@ -4,6 +4,7 @@ using SME.SGP.Aplicacao;
 using SME.SGP.Aplicacao.Integracoes;
 using SME.SGP.Dominio.Interfaces;
 using SME.SGP.Infra;
+using SME.SGP.Infra.Dtos;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -123,7 +124,10 @@ namespace SME.SGP.Dominio.Servicos
                 await mediator.Send(new IncluirFilaCalcularFrequenciaPorTurmaCommand(codigosAlunosCompensacao, periodo.PeriodoFim, compensacaoDto.TurmaId, compensacaoDto.DisciplinaId));
             }
 
-            Cliente.Executar<IServicoNotificacaoFrequencia>(c => c.NotificarCompensacaoAusencia(compensacao.Id));
+            await mediator.Send(new PublicarFilaSgpCommand(RotasRabbitSgp.NotificarCompensacaoAusencia,
+                                                           new FiltroNotificacaoCompensacaoAusenciaDto(compensacao.Id),
+                                                           Guid.NewGuid(),
+                                                           usuario));
         }
         private async Task MoverRemoverExcluidos(string novo, string atual)
         {
