+++ src/SME.SGP.Dominio.Servicos/ServicoNotificacaoFrequencia.cs
@@ -16,24 +16,18 @@ namespace SME.SGP.Dominio.Servicos
     public class ServicoNotificacaoFrequencia : IServicoNotificacaoFrequencia
     {
         private readonly IConfiguration configuration;
-        private readonly IRepositorioCompensacaoAusencia repositorioCompensacaoAusencia;
-        private readonly IRepositorioCompensacaoAusenciaAluno repositorioCompensacaoAusenciaAluno;
-        private readonly IRepositorioDre repositorioDre;
         private readonly IRepositorioFrequencia repositorioFrequencia;
         private readonly IRepositorioFrequenciaAlunoDisciplinaPeriodo repositorioFrequenciaAluno;
-        private readonly IRepositorioNotificacaoCompensacaoAusencia repositorioNotificacaoCompensacaoAusencia;
         private readonly IRepositorioNotificacaoFrequencia repositorioNotificacaoFrequencia;
         private readonly IRepositorioParametrosSistema repositorioParametrosSistema;
         private readonly IRepositorioPeriodoEscolar repositorioPeriodoEscolar;
         private readonly IRepositorioTipoCalendario repositorioTipoCalendario;
         private readonly IRepositorioComponenteCurricular repositorioComponenteCurricular;
         private readonly IRepositorioTurma repositorioTurma;
-        private readonly IRepositorioUe repositorioUe;
         private readonly IServicoEol servicoEOL;
         private readonly IServicoNotificacao servicoNotificacao;
         private readonly IServicoUsuario servicoUsuario;
         private readonly IConsultasFeriadoCalendario consultasFeriadoCalendario;
-        private readonly IMediator mediator;
 
 
         public ServicoNotificacaoFrequencia(IRepositorioNotificacaoFrequencia repositorioNotificacaoFrequencia,
@@ -41,20 +35,13 @@ namespace SME.SGP.Dominio.Servicos
                                             IRepositorioFrequencia repositorioFrequencia,
                                             IRepositorioComponenteCurricular repositorioComponenteCurricular,
                                             IRepositorioFrequenciaAlunoDisciplinaPeriodo repositorioFrequenciaAluno,
-                                            IRepositorioCompensacaoAusencia repositorioCompensacaoAusencia,
-                                            IRepositorioCompensacaoAusenciaAluno repositorioCompensacaoAusenciaAluno,
                                             IRepositorioTurma repositorioTurma,
-                                            IRepositorioUe repositorioUe,
-                                            IRepositorioDre repositorioDre,
-                                            IRepositorioNotificacaoCompensacaoAusencia repositorioNotificacaoCompensacaoAusencia,
                                             IRepositorioPeriodoEscolar repositorioPeriodoEscolar,
                                             IRepositorioTipoCalendario repositorioTipoCalendario,
                                             IServicoNotificacao servicoNotificacao,
                                             IServicoUsuario servicoUsuario,
                                             IServicoEol servicoEOL,
-                                            IConfiguration configuration,
-                                            IConsultasFeriadoCalendario consultasFeriadoCalendario,
-                                            IMediator mediator)
+                                            IConfiguration configuration)
         {
             this.repositorioNotificacaoFrequencia = repositorioNotificacaoFrequencia ?? throw new ArgumentNullException(nameof(repositorioNotificacaoFrequencia));
             this.repositorioParametrosSistema = repositorioParametrosSistema ?? throw new ArgumentNullException(nameof(repositorioParametrosSistema));
@@ -62,19 +49,13 @@ namespace SME.SGP.Dominio.Servicos
             this.repositorioFrequencia = repositorioFrequencia ?? throw new ArgumentNullException(nameof(repositorioFrequencia));
             this.repositorioFrequenciaAluno = repositorioFrequenciaAluno ?? throw new ArgumentNullException(nameof(repositorioFrequenciaAluno));
             this.servicoUsuario = servicoUsuario ?? throw new ArgumentNullException(nameof(servicoUsuario));
-            this.repositorioCompensacaoAusencia = repositorioCompensacaoAusencia ?? throw new ArgumentNullException(nameof(repositorioCompensacaoAusencia));
-            this.repositorioCompensacaoAusenciaAluno = repositorioCompensacaoAusenciaAluno ?? throw new ArgumentNullException(nameof(repositorioCompensacaoAusenciaAluno));
             this.repositorioTurma = repositorioTurma ?? throw new ArgumentNullException(nameof(repositorioTurma));
-            this.repositorioUe = repositorioUe ?? throw new ArgumentNullException(nameof(repositorioUe));
-            this.repositorioDre = repositorioDre ?? throw new ArgumentNullException(nameof(repositorioDre));
-            this.repositorioNotificacaoCompensacaoAusencia = repositorioNotificacaoCompensacaoAusencia ?? throw new ArgumentNullException(nameof(repositorioNotificacaoCompensacaoAusencia));
             this.repositorioPeriodoEscolar = repositorioPeriodoEscolar ?? throw new ArgumentNullException(nameof(repositorioPeriodoEscolar));
             this.repositorioTipoCalendario = repositorioTipoCalendario ?? throw new ArgumentNullException(nameof(repositorioTipoCalendario));
             this.servicoEOL = servicoEOL ?? throw new ArgumentNullException(nameof(servicoEOL));
             this.configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
             this.repositorioComponenteCurricular = repositorioComponenteCurricular ?? throw new ArgumentNullException(nameof(repositorioComponenteCurricular));
             this.consultasFeriadoCalendario = consultasFeriadoCalendario ?? throw new System.ArgumentNullException(nameof(consultasFeriadoCalendario));
-            this.mediator = mediator ?? throw new System.ArgumentNullException(nameof(mediator));
         }
 
         #region Metodos Publicos
@@ -110,94 +91,6 @@ namespace SME.SGP.Dominio.Servicos
 
             await NotificaAlunosFaltososCargo(DiaRetroativo(dataReferencia, quantidadeDiasCP - 1), quantidadeDiasCP, Cargo.CP, tipoCalendario?.Id ?? 0);
             //await NotificaAlunosFaltososCargo(DiaRetroativo(dataReferencia, quantidadeDiasDiretor - 1), quantidadeDiasDiretor, Cargo.Diretor, tipoCalendario?.Id ?? 0);
-        }
-
-        public async Task NotificarCompensacaoAusencia(long compensacaoId)
-        {
-            // Verifica se compensação possui alunos vinculados
-            var alunos = await repositorioCompensacaoAusenciaAluno.ObterPorCompensacao(compensacaoId);
-            if (alunos == null || !alunos.Any())
-                return;
-
-            // Verifica se possui aluno não notificado na compensação
-            if (!alunos.Any(a => !a.Notificado && a.QuantidadeFaltasCompensadas > 0))
-                return;
-
-            // Carrega dados da compensacao a notificar
-            var compensacao = repositorioCompensacaoAusencia.ObterPorId(compensacaoId);
-
-            var turma = await repositorioTurma.ObterTurmaComUeEDrePorId(compensacao.TurmaId);
-
-            var disciplinaEOL = await ObterNomeDisciplina(compensacao.DisciplinaId);
-
-            MeusDadosDto professor = await servicoEOL.ObterMeusDados(compensacao.CriadoRF);
-
-            var possuirPeriodoAberto = await mediator.Send(new TurmaEmPeriodoAbertoQuery(turma, DateTimeExtension.HorarioBrasilia(), compensacao.Bimestre, true));
-            var parametroAtivo = await mediator.Send(new ObterParametroSistemaPorTipoEAnoQuery(TipoParametroSistema.PermiteCompensacaoForaPeriodo, turma.AnoLetivo));
-
-            // Carrega dados dos alunos não notificados
-            var alunosTurma = await servicoEOL.ObterAlunosPorTurma(turma.CodigoTurma);
-            var alunosDto = new List<CompensacaoAusenciaAlunoQtdDto>();
-            foreach (var aluno in alunos)
-            {
-                var alunoEol = alunosTurma.FirstOrDefault(a => a.CodigoAluno == aluno.CodigoAluno);
-                alunosDto.Add(new CompensacaoAusenciaAlunoQtdDto()
-                {
-                    NumeroAluno = alunoEol.NumeroAlunoChamada,
-                    CodigoAluno = aluno.CodigoAluno,
-                    NomeAluno = alunoEol.NomeAluno,
-                    QuantidadeCompensacoes = aluno.QuantidadeFaltasCompensadas
-                });
-            }
-
-            repositorioNotificacaoCompensacaoAusencia.Excluir(compensacaoId);
-
-            var cargos = new Cargo[] { Cargo.CP };
-            if (GerarNotificacaoExtemporanea(possuirPeriodoAberto, parametroAtivo != null ? parametroAtivo.Ativo : false))
-            {
-
-                await NotificarCompensacaoExtemporanea(
-                     professor.Nome,
-                     professor.CodigoRf,
-                     disciplinaEOL,
-                     turma.CodigoTurma,
-                     turma.Nome,
-                     turma.ModalidadeCodigo.ObterNomeCurto(),
-                     turma.Ue.CodigoUe,
-                     turma.Ue.Nome,
-                     turma.Ue.TipoEscola.ObterNomeCurto(),
-                     turma.Ue.Dre.CodigoDre,
-                     turma.Ue.Dre.Nome,
-                     compensacao.Bimestre,
-                     compensacao.Nome,
-                     alunosDto, cargos);
-            }
-            else
-            {
-                await NotificarCompensacaoAusencia(
-                         professor.Nome
-                        , professor.CodigoRf
-                        , disciplinaEOL
-                        , turma.CodigoTurma
-                        , turma.Nome
-                        , turma.ModalidadeCodigo.ObterNomeCurto()
-                        , turma.Ue.CodigoUe
-                        , turma.Ue.Nome
-                        , turma.Ue.TipoEscola.ObterNomeCurto()
-                        , turma.Ue.Dre.CodigoDre
-                        , turma.Ue.Dre.Nome
-                        , compensacao.Bimestre
-                        , compensacao.Nome
-                        , alunosDto, cargos);
-            }
-            // Marca aluno como notificado
-            alunosDto.ForEach(alunoDto =>
-            {
-                var aluno = alunos.FirstOrDefault(a => a.CodigoAluno == alunoDto.CodigoAluno);
-                aluno.Notificado = true;
-                repositorioCompensacaoAusenciaAluno.Salvar(aluno);
-            });
-
         }
 
         public async Task VerificaRegraAlteracaoFrequencia(long registroFrequenciaId, DateTime criadoEm, DateTime alteradoEm, long usuarioAlteracaoId)
@@ -612,36 +505,6 @@ namespace SME.SGP.Dominio.Servicos
             return cargosNotificados;
         }
 
-        private async Task<long> NotificarCompensacaoAusencia(string professor, string professorRf, string disciplina,
-            string codigoTurma, string turma, string modalidade, string codigoUe, string escola, string tipoEscola, string codigoDre, string dre,
-            int bimestre, string atividade, List<CompensacaoAusenciaAlunoQtdDto> alunos, Cargo[] cargos)
-        {
-            var tituloMensagem = $"Atividade de compensação da turma {turma}";
-
-            StringBuilder mensagemUsuario = new StringBuilder();
-            mensagemUsuario.AppendLine($"<p>A atividade de compensação <b>'{atividade}'</b> do componente curricular de <b>{disciplina}</b> foi cadastrada para a turma <b>{turma} {modalidade}</b> da <b>{tipoEscola} {escola} ({dre})</b> no <b>{bimestre}º</b> Bimestre pelo professor <b>{professor} ({professorRf})</b>.</p>");
-            mensagemUsuario.AppendLine("<p>O(s) seguinte(s) aluno(s) foi(ram) vinculado(s) a atividade:</p>");
-
-            mensagemUsuario.AppendLine("<table style='margin-left: auto; margin-right: auto;' border='2' cellpadding='5'>");
-            mensagemUsuario.AppendLine("<tr>");
-            mensagemUsuario.AppendLine("<td style='padding: 5px;'>Nº</td>");
-            mensagemUsuario.AppendLine("<td style='padding: 5px;'>Nome do aluno</td>");
-            mensagemUsuario.AppendLine("<td style='padding: 5px;'>Quantidade de aulas compensadas</td>");
-            mensagemUsuario.AppendLine("</tr>");
-            foreach (var aluno in alunos)
-            {
-                mensagemUsuario.AppendLine("<tr>");
-                mensagemUsuario.Append($"<td style='padding: 5px;'>{aluno.NumeroAluno}</td>");
-                mensagemUsuario.Append($"<td style='padding: 5px;'>{aluno.NomeAluno}</td>");
-                mensagemUsuario.Append($"<td style='text-align: center;'>{aluno.QuantidadeCompensacoes}</td>");
-                mensagemUsuario.AppendLine("</tr>");
-            }
-            mensagemUsuario.AppendLine("</table>");
-            mensagemUsuario.Append("Para consultar os detalhes desta atividade acesse 'Diário de classe > Compensação de ausência'");
-
-            return await mediator.Send(new EnviarNotificacaoCommand(tituloMensagem, mensagemUsuario.ToString(), NotificacaoCategoria.Aviso, NotificacaoTipo.Frequencia, cargos, codigoDre, codigoUe, codigoTurma));
-        }
-
         private async Task NotificaRegistroFrequencia(Usuario usuario, RegistroFrequenciaFaltanteDto turmaSemRegistro, TipoNotificacaoFrequencia tipo)
         {
             var disciplinas = await repositorioComponenteCurricular.ObterDisciplinasPorIds(new long[] { long.Parse(turmaSemRegistro.DisciplinaId) });
@@ -746,19 +609,6 @@ namespace SME.SGP.Dominio.Servicos
 
             return dataRetorno;
         }
-        private bool GerarNotificacaoExtemporanea(bool periodoAberto, bool parametroAtivo)
-        {
-            if (periodoAberto)
-                return false;
-            else if (parametroAtivo && !periodoAberto)
-                return true;
-            else if (!parametroAtivo && !periodoAberto)
-                throw new NegocioException("Compensação de ausência não permitida, É necessário que o período esteja aberto");
-            else if (!parametroAtivo)
-                return false;
-
-            return false;
-        }
         private bool Feriado(DateTime data)
         {
             FiltroFeriadoCalendarioDto filtro = new FiltroFeriadoCalendarioDto();
@@ -766,34 +616,6 @@ namespace SME.SGP.Dominio.Servicos
             var ret = consultasFeriadoCalendario.Listar(filtro).Result;
             return ret.Any(x => x.DataFeriado == data);
         }
-        private async Task<long> NotificarCompensacaoExtemporanea(string professor, string professorRf, string disciplina, string codigoTurma, string turma, string modalidade, string codigoUe, string escola, string tipoEscola, string codigoDre, string dre, int bimestre, string atividade, List<CompensacaoAusenciaAlunoQtdDto> alunos, Cargo[] cargos)
-        {
-            var tituloMensagem = $"Atividade de compensação de ausência extemporânea - {modalidade}-{turma} - {disciplina}";
-
-            StringBuilder mensagemUsuario = new StringBuilder();
-            mensagemUsuario.AppendLine($"<p>A atividade de compensação <b>'{atividade}'</b> do componente curricular de <b>{disciplina}</b> foi cadastrada para a turma <b>{turma} {modalidade}</b> da <b>{tipoEscola} {escola} ({dre})</b> no <b>{bimestre}º</b> Bimestre pelo professor <b>{professor} ({professorRf})</b> de forma extemporânea (fora do período escolar).</p>");
-            mensagemUsuario.AppendLine("<p>O(s) seguinte(s) aluno(s) foi(ram) vinculado(s) a atividade:</p>");
-
-            mensagemUsuario.AppendLine("<table style='margin-left: auto; margin-right: auto;' border='2' cellpadding='5'>");
-            mensagemUsuario.AppendLine("<tr>");
-            mensagemUsuario.AppendLine("<td style='padding: 5px;'>Nº</td>");
-            mensagemUsuario.AppendLine("<td style='padding: 5px;'>Nome do aluno</td>");
-            mensagemUsuario.AppendLine("<td style='padding: 5px;'>Quantidade de aulas compensadas</td>");
-            mensagemUsuario.AppendLine("</tr>");
-            foreach (var aluno in alunos)
-            {
-                mensagemUsuario.AppendLine("<tr>");
-                mensagemUsuario.Append($"<td style='padding: 5px;'>{aluno.NumeroAluno}</td>");
-                mensagemUsuario.Append($"<td style='padding: 5px;'>{aluno.NomeAluno}</td>");
-                mensagemUsuario.Append($"<td style='text-align: center;'>{aluno.QuantidadeCompensacoes}</td>");
-                mensagemUsuario.AppendLine("</tr>");
-            }
-            mensagemUsuario.AppendLine("</table>");
-
-            return await mediator.Send(new EnviarNotificacaoCommand(tituloMensagem, mensagemUsuario.ToString(), NotificacaoCategoria.Alerta, NotificacaoTipo.Frequencia, cargos, codigoDre, codigoUe, codigoTurma));
-
-        }
-
 
         #endregion Metodos Privados
     }
