+++ src/SME.SGP.Infra/Dtos/CompensacaoAusencia/FiltroNotificacaoCompensacaoAusenciaDto.cs
@@ -1,0 +1,12 @@
+﻿namespace SME.SGP.Infra.Dtos
+{
+    public class FiltroNotificacaoCompensacaoAusenciaDto 
+    {
+        public FiltroNotificacaoCompensacaoAusenciaDto(long compensacaoId)
+        {
+            CompensacaoId = compensacaoId;
+        }
+
+        public long CompensacaoId { get; set; }
+    }
+}
