+++ src/SME.SGP.Infra/Fila/RotasRabbitSgp.cs
@@ -83,6 +83,8 @@
         public const string SincronizaEstruturaInstitucionalCicloSync = "sgp.sincronizacao.institucional.ciclo.sync";
         public const string SincronizaEstruturaInstitucionalCicloTratar = "sgp.sincronizacao.institucional.ciclo.tratar";
 
+        public const string NotificarCompensacaoAusencia = "sgp.compensacao.ausencia.notificar";
+
         public const string ConsolidacaoFrequenciasTurmasCarregar = "sgp.frequencia.turma.carregar";
         public const string ConsolidarFrequenciasTurmasNoAno = "sgp.frequencia.turma.ano.consolidar";
         public const string ConsolidarFrequenciasPorTurma = "sgp.frequencia.turma.consolidar";
